Github Actions: Update Node.js version in workflows
Its better to let github actions use .nvmrc to determine
the node version to use in the workflows. This
helps with ensuring that github actions are using the same node
version as the local development environment.

As you can see already the node versions were different
from nvmrc which is specifically using 22.11.0