Update "Check Your Land Actions" page to support one or more parcels
The "check your land actions" page has been updated to support
one or more parcels. This change allows users to view the selected
land actions for multiple parcels simultaneously.

Another change is that the "Check Your Land Actions" controller now
returns the page title instead of all the logic in the view to plurialize
parcel/action and display the correct total selected parcels/actions. This
change also makes it easier to test than trying to render the view with
data in different states.