Change "Select actions" page form tabular form to conditionally revealed content
This commit changes the checkboxes in the table to standard govuk checkboxes
component with conditionally revealed content. The actions quantities are
only shown when the checkbox is checked, and the "Select actions" page
is now more accessible.

It should also now be possible to switch between parcels and the actions
selected for each parcel without losing the selected actions or quantities.

This commit does not cover fixing the styling of error messages.