Validate data at UI and API level in the same call
Previously we were going simple "not null" validations in the UI and if any fields failed then we would not validate the rest of the supplied data. The user would then have to correct that, submit it only to get a list of API more complicated validation errors messages. This is not a great user experience as the user might assume the rest of the data was successful and the service leaves them feeling like they are chasing validation error messages.

With these changes, we build up a new object containing only actions that do have values. this new object is passed to the API to validate. If any validations fail we concatenate those error messages with any UI validation messages. This is much better user experience by removing unnecessary form submissions.