Add support for organisation names that contain colons
Unfortunately Defra ID does not escape organisation names that have colons. What we do know is that there is a data structure for the relationship so if we take the first couple parts and the last couple parts, remove those then we should be left with the organisation name regardless as to whether it has one or more colons in its name.