Remove inline error message
We removing this error message as we have two "groupings" of radio buttons which we would want to show as being invalid. The GOV.UK Design System doesn't support what we want to do on this page. We also know the design of this page will change to help support farmers select multiple land actions/land parcels.